refactor(documents): fetch the review where the reader asks for it
Opening a review is one request, triggered by a click, used once. It was a
cached query with every cache feature switched off: a key made unique per open,
a zero collection time, a manual enabled flag, and a placeholder key for when
there was nothing to fetch. That is a caching layer configured never to cache.

Ask for the verdict in showReview and keep the document it was computed against
alongside the review itself. Staleness stops being a question rather than being
answered again: there is no second copy to go out of date.

Deletes the query module. The overlay is now a hook that shows marks when a
review names its document.
