+++ src/features/workspaces/documents/document-edit-review-context.tsx
@@ -1,24 +1,30 @@
 import { createContext, type ReactNode, use, useCallback, useMemo, useState } from "react";
+import { toast } from "sonner";
 
+import type { DocumentEditReceiptUnavailableStatus } from "#/features/workspaces/documents/document-edit-receipt";
+import { getDocumentEditReceiptReviewFn } from "#/features/workspaces/documents/document-edit-review-functions";
+import type { TiptapDocumentJson } from "#/features/workspaces/documents/tiptap-document";
 import { useWorkspaceLocationActions } from "#/features/workspaces/locations/workspace-location-context";
 
 /**
  * Review belongs to the document, not to the view showing it. Tying it to a
  * view instance meant the review had to be opened after that view existed, and
  * every wherever-it-is-now question became a lifecycle problem.
+ *
+ * The document it was computed against travels with it. Whether an edit can
+ * still be reviewed is a judgement about a moment, so it is asked once, when
+ * the reader asks to see it, and never revisited behind their back.
  */
 export interface ActiveDocumentEditReview {
+	beforeDocument: TiptapDocumentJson;
 	itemId: string;
-	/** When this review was opened. The overlay ignores any verdict older than
-	 * this, because a cached one describes a document that has since moved on. */
-	openedAt: number;
 	receiptIds: string[];
 }
 
 interface DocumentEditReviewContextValue {
 	activeReview: ActiveDocumentEditReview | null;
 	hideReview: () => void;
-	showReview: (input: { itemId: string; receiptIds: string[] }) => boolean;
+	showReview: (input: { itemId: string; receiptIds: string[] }) => Promise<void>;
 	workspaceId: string;
 }
 
@@ -35,21 +41,34 @@ export function DocumentEditReviewProvider({
 	const [activeReview, setActiveReview] = useState<ActiveDocumentEditReview | null>(null);
 	const hideReview = useCallback(() => setActiveReview(null), []);
 	const showReview = useCallback(
-		(input: { itemId: string; receiptIds: string[] }) => {
+		async (input: { itemId: string; receiptIds: string[] }) => {
 			// reveal opens the document, or focuses the tab already holding it, and
 			// only fails when the item is gone.
 			if (!reveal({ itemId: input.itemId, kind: "item", version: 1 })) {
-				return false;
+				toast.error("This document no longer exists.");
+				return;
+			}
+
+			const review = await getDocumentEditReceiptReviewFn({
+				data: { itemId: input.itemId, receiptIds: input.receiptIds, workspaceId },
+			}).catch(() => null);
+
+			if (!review) {
+				toast.error("Could not load these changes.");
+				return;
+			}
+			if (review.status !== "ready") {
+				toast.error(unavailableReviewMessages[review.status]);
+				return;
 			}
 
 			setActiveReview({
+				beforeDocument: review.beforeDocument,
 				itemId: input.itemId,
-				openedAt: Date.now(),
 				receiptIds: input.receiptIds,
 			});
-			return true;
 		},
-		[reveal],
+		[reveal, workspaceId],
 	);
 	const value = useMemo(
 		() => ({ activeReview, hideReview, showReview, workspaceId }),
@@ -67,3 +86,11 @@ export function useDocumentEditReview() {
 
 	return value;
 }
+
+const unavailableReviewMessages: Record<DocumentEditReceiptUnavailableStatus, string> = {
+	content_changed: "The document changed after this AI edit.",
+	not_found: "These changes are no longer available.",
+	not_latest: "Only the latest unchanged AI edit can be reviewed.",
+	reverted: "These changes were already undone.",
+	review_unavailable: "Change review is unavailable for this large document.",
+};
